Remove access check when asking for location
We don’t need to check access when asking for location for dataset. This will be done when we need key for read/write

TODO: Need to handle when there is no dataset on read. We plan to use String for version in to support this https://github.com/statisticsnorway/data-access-protobuf/blob/master/src/main/proto/service.proto